+++ packages/apps/dev-wallet/src/pages/create-profile/create-profile.tsx
@@ -4,7 +4,7 @@ import { LayoutContext } from '@/modules/layout/layout.provider';
 import { useNetwork } from '@/modules/network/network.hook';
 import { IKeySource } from '@/modules/wallet/wallet.repository';
 import { kadenaGenMnemonic } from '@kadena/hd-wallet';
-import { MonoCheck } from '@kadena/react-icons/system';
+import { MonoCheck } from '@kadena/kode-icons/system';
 import { Button, Heading, Stack, Text, TextField } from '@kadena/kode-ui';
 import { useContext, useState } from 'react';
 import { Controller, useForm } from 'react-hook-form';