+++ packages/apps/docs/src/components/Markdown/MDNotification/utils.tsx
@@ -1,5 +1,5 @@
+import { MonoWarning } from '@kadena/kode-icons';
 import type { INotificationProps } from '@kadena/kode-ui';
-import { MonoWarning } from '@kadena/react-icons';
 import React from 'react';
 
 export type LabelType =