+++ packages/apps/graph-client/src/components/token-table/token-table.tsx
@@ -1,4 +1,5 @@
 import type { NonFungibleTokenBalance } from '@/__generated__/sdk';
+import { KSquareKdacolorGreen } from '@kadena/kode-icons/brand';
 import {
   Box,
   Cell,
@@ -10,7 +11,6 @@ import {
   TableHeader,
 } from '@kadena/kode-ui';
 import { atoms } from '@kadena/kode-ui/styles';
-import { KSquareKdacolorGreen } from '@kadena/react-icons/brand';
 import React from 'react';
 
 interface ITokenTableProps {