+++ packages/apps/tools/src/components/Global/RequestKeyField/index.tsx
@@ -1,6 +1,6 @@
+import { MonoKey } from '@kadena/kode-icons/system';
 import type { ITextFieldProps } from '@kadena/kode-ui';
 import { TextField } from '@kadena/kode-ui';
-import { MonoKey } from '@kadena/react-icons/system';
 import useTranslation from 'next-translate/useTranslation';
 import type { FC } from 'react';
 import React, { forwardRef } from 'react';