+++ packages/apps/tools/src/components/Partials/transactions/transfer/right-info-sidebar.tsx
@@ -10,9 +10,9 @@ import {
 } from './styles.css';
 
 import { useIsMatchingMediaQuery } from '@/hooks/use-is-mobile-media-query';
+import { MonoInfo, MonoLink } from '@kadena/kode-icons/system';
 import { Accordion, AccordionItem } from '@kadena/kode-ui';
 import { breakpoints } from '@kadena/kode-ui/styles';
-import { MonoInfo, MonoLink } from '@kadena/react-icons/system';
 import useTranslation from 'next-translate/useTranslation';
 import { useRouter } from 'next/router';
 import type { FC } from 'react';