+++ packages/apps/tools/src/components/Partials/transactions/transfer/sign-form.tsx
@@ -24,10 +24,7 @@ import { useWalletConnectClient } from '@/context/connect-wallet-context';
 
 import type { AccountDetails } from '@/hooks/use-account-details-query';
 import { stripAccountPrefix } from '@/utils/string';
-import {
-  MonoKeyboardArrowRight,
-  MonoRefresh,
-} from '@kadena/react-icons/system';
+import { MonoKeyboardArrowRight, MonoRefresh } from '@kadena/kode-icons/system';
 import type { ChainId } from '@kadena/types';
 import type { PactCommandObject } from '@ledgerhq/hw-app-kda';
 import { z } from 'zod';