+++ packages/libs/kode-ui/src/components/Breadcrumbs/Breadcrumbs.test.tsx
@@ -1,4 +1,4 @@
-import { MonoKAccount } from '@kadena/react-icons/system';
+import { MonoKAccount } from '@kadena/kode-icons/system';
 import { render } from '@testing-library/react';
 import React from 'react';
 import { describe, expect, test } from 'vitest';