+++ packages/libs/kode-ui/src/components/Form/Select/Select.stories.tsx
@@ -1,4 +1,4 @@
-import { MonoAccountCircle, MonoAdd } from '@kadena/react-icons/system';
+import { MonoAccountCircle, MonoAdd } from '@kadena/kode-icons/system';
 import type { Meta, StoryObj } from '@storybook/react';
 import React, { useState } from 'react';
 import {